requirements_old.txt
Package                       Version     Editable project location
----------------------------- ----------- ----------------------------------------------
actionlib                     1.12.1
aiohappyeyeballs              2.4.4
aiohttp                       3.10.11
aiosignal                     1.3.1
angles                        1.9.12
annotated-types               0.7.0
async-timeout                 5.0.1
attrs                         25.1.0
base_local_planner            1.16.7
bondpy                        1.8.5
camera_calibration            1.15.2
camera_calibration_parsers    1.11.13
catkin                        0.7.29
catkin-pkg                    1.0.0
certifi                       2025.1.31
charset-normalizer            3.4.1
clarabel                      0.9.0
contourpy                     1.1.1
controller_manager            0.18.4
controller_manager_msgs       0.18.4
cv_bridge                     1.13.1
cvxpy                         1.5.2
cycler                        0.12.1
diagnostic_analysis           1.9.7
diagnostic_common_diagnostics 1.9.7
diagnostic_updater            1.9.7
dill                          0.4.0
distro                        1.9.0
docutils                      0.20.1
dynamic_reconfigure           1.6.5
ecos                          2.0.14
fonttools                     4.56.0
frozenlist                    1.5.0
gazebo_plugins                2.8.7
gazebo_ros                    2.8.7
gencpp                        0.6.6
geneus                        2.2.6
genlisp                       0.4.16
genmsg                        0.5.17
gennodejs                     2.0.1
genpy                         0.6.16
greenlet                      3.1.1
idna                          3.10
image_geometry                1.13.1
importlib_resources           6.4.5
interactive-markers           1.11.5
joblib                        1.4.2
joint_state_publisher         1.12.15
kdl-parser-py                 1.13.3
kiwisolver                    1.4.7
laser_geometry                1.6.7
matplotlib                    3.7.5
message_filters               1.14.13
multidict                     6.1.0
numpy                         1.24.4
openai                        0.28.0
osqp                          0.6.7.post3
packaging                     24.2
pillow                        10.4.0
pip                           25.0.1
pkg_resources                 0.0.0
propcache                     0.2.0
pydantic                      2.10.6
pydantic_core                 2.27.2
pymap3d                       3.1.0
pyparsing                     3.1.4
python-dateutil               2.9.0.post0
python_qt_binding             0.4.4
PyYAML                        6.0.2
pyzonotope                    1.0.0       /home/jetson/catkin_ws/src/zonotope/pyzonotope
qdldl                         0.1.7.post5
qt-dotgraph                   0.4.2
qt-gui                        0.4.2
qt-gui-cpp                    0.4.2
qt-gui-py-common              0.4.2
requests                      2.32.3
resource_retriever            1.12.7
rosbag                        1.14.13
rosboost-cfg                  1.14.9
rosclean                      1.14.9
roscreate                     1.14.9
rosdep                        0.25.1
rosdistro                     1.0.1
rosgraph                      1.14.13
rosinstall                    0.7.8
rosinstall-generator          0.1.23
roslaunch                     1.14.13
roslib                        1.14.9
roslint                       0.11.2
roslz4                        1.14.13
rosmake                       1.14.9
rosmaster                     1.14.13
rosmsg                        1.14.13
rosnode                       1.14.13
rosparam                      1.14.13
rospkg                        1.6.0
rospy                         1.14.13
rosservice                    1.14.13
rostest                       1.14.13
rostopic                      1.14.13
rosunit                       1.14.9
roswtf                        1.14.13
rqt_action                    0.4.9
rqt_bag                       0.5.1
rqt_bag_plugins               0.5.1
rqt_console                   0.4.9
rqt_dep                       0.4.9
rqt_graph                     0.4.11
rqt_gui                       0.5.3
rqt_gui_py                    0.5.3
rqt-image-view                0.4.17
rqt_launch                    0.4.8
rqt_logger_level              0.4.8
rqt-moveit                    0.5.10
rqt_msg                       0.4.8
rqt_nav_view                  0.5.7
rqt_plot                      0.4.13
rqt_pose_view                 0.5.8
rqt_publisher                 0.4.8
rqt_py_common                 0.5.3
rqt_py_console                0.4.8
rqt-reconfigure               0.5.4
rqt_robot_dashboard           0.5.7
rqt-robot-monitor             0.5.14
rqt_robot_steering            0.5.10
rqt_runtime_monitor           0.5.7
rqt-rviz                      0.7.0
rqt_service_caller            0.4.8
rqt_shell                     0.4.9
rqt_srv                       0.4.8
rqt_tf_tree                   0.6.0
rqt_top                       0.4.8
rqt_topic                     0.4.11
rqt_web                       0.4.8
rviz                          1.13.30
scikit-learn                  1.3.2
scipy                         1.10.1
scs                           3.2.7.post2
sensor-msgs                   1.12.8
setuptools                    75.3.0
shapely                       2.0.7
six                           1.17.0
smach                         2.0.3
smach_ros                     2.0.3
smclib                        1.8.5
snakeviz                      2.2.1
sound-play                    0.3.17
SQLAlchemy                    2.0.41
stl_constrained_zpc           0.1.3       /home/jetson/catkin_ws/src/stl_constrained_zpc
tf                            1.12.1
tf_conversions                1.12.1
tf2_geometry_msgs             0.6.5
tf2_kdl                       0.6.5
tf2_py                        0.6.5
tf2_ros                       0.6.5
threadpoolctl                 3.5.0
topic_tools                   1.14.13
tornado                       6.4.2
tqdm                          4.67.1
transforms3d                  0.4.2
typing_extensions             4.12.2
urdfdom-py                    0.4.6
urllib3                       2.2.3
vcstools                      0.1.42
websockets                    13.1
wheel                         0.45.1
wstool                        0.1.17
xacro                         1.13.20
yarl                          1.15.2
zipp                          3.20.2